refactor: extract sync_adapters into astra-sync-adapters crate (1487 LOC)
Moves runtime/src/sync_adapters.rs into a new astra-sync-adapters crate.
The module bridges learning modules (EntityGraph, PatternLibrary, Calibrator,
ToolHealth) with the unified sync engine's DomainAdapter trait — a natural
standalone unit now that astra-evolution::persistence is available.

Stacks on top of the pipeline/persistence extraction (PR #124).

- New crate depends on astra-evolution, astra-pipeline, astra-services, astra-text-utils
- runtime/src/sync_adapters.rs is now a 6-line re-export stub
- Cargo.toml workspace updated; runtime adds astra-sync-adapters dep
- No behavior changes; all paths resolve via re-exports
