+++ rust/crates/astra-sync-adapters/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "astra-sync-adapters"
+version.workspace = true
+edition.workspace = true
+description = "Cloud sync adapters bridging learning modules with the unified sync engine"
+
+[dependencies]
+async-trait = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+
+# Internal crates
+astra-evolution = { path = "../astra-evolution" }
+astra-pipeline = { path = "../astra-pipeline" }
+astra-services = { path = "../services" }
+astra-text-utils = { path = "../astra-text-utils" }
+
+[dev-dependencies]
+tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
