+++ rust/crates/runtime/Cargo.toml
@@ -43,6 +43,7 @@ astra-prompts.workspace = true
 astra-sandbox.workspace = true
 astra-server-types.workspace = true
 astra-skills.workspace = true
+astra-sync-adapters.workspace = true
 astra-turn-core.workspace = true
 astra-turn-types.workspace = true
 astra-services.workspace = true
